refactor: moving some furniture around
